refactor: move member adding to manager
